First try at using the set_protocol directly. It has side effects on the closing of the comm channel, needs work. Tests fail currently
