+++ roll/__init__.py
@@ -8,18 +8,23 @@ If you do not understand why something is not working as expected,
 please submit an issue (or even better a pull-request with at least
 a test failing): https://github.com/pyrates/roll/issues/new
 """
+
 import asyncio
-from autoroutes import Routes
-from biscuits import Cookie, parse
+
 from collections import namedtuple
 from http import HTTPStatus
 from io import BytesIO
-from multifruits import Parser, extract_filename, parse_content_disposition
-from typing import TypeVar
-from urllib.parse import parse_qs
-from .protocols import Protocol, WSProtocol, ConnectionClosed
 from traceback import print_exc
+from typing import TypeVar
+from urllib.parse import unquote, parse_qs
 
+from autoroutes import Routes
+from biscuits import Cookie, parse
+from httptools import (
+    HttpParserUpgrade, HttpParserError, HttpRequestParser, parse_url)
+from multifruits import Parser, extract_filename, parse_content_disposition
+
+from .websocket import WebsocketHandler
 
 try:
     # In case you use json heavily, we recommend installing
@@ -304,11 +309,101 @@ class Response:
             self._cookies = self.app.Cookies()
         return self._cookies
 
-            
+
+class Protocol(asyncio.Protocol):
+    """Responsible of parsing the request and writing the response."""
+
+    __slots__ = ('app', 'request', 'parser', 'response', 'writer', 'upgrade')
+    _BODYLESS_METHODS = ('HEAD', 'CONNECT')
+    _BODYLESS_STATUSES = (HTTPStatus.CONTINUE, HTTPStatus.SWITCHING_PROTOCOLS,
+                          HTTPStatus.PROCESSING, HTTPStatus.NO_CONTENT,
+                          HTTPStatus.NOT_MODIFIED)
+    RequestParser = HttpRequestParser
+
+    def __init__(self, app):
+        self.app = app
+        self.parser = self.RequestParser(self)
+        self.upgrade = None
+
+    def connection_made(self, transport):
+        self.writer = transport
+        
+    def data_received(self, data: bytes):
+        try:
+            self.parser.feed_data(data)
+        except HttpParserError:
+            # If the parsing failed before on_message_begin, we don't have a
+            # response.
+            self.response = self.app.Response(self.app)
+            self.response.status = HTTPStatus.BAD_REQUEST
+            self.response.body = b'Unparsable request'
+            self.write()
+        except HttpParserUpgrade:
+            self.upgrade = self.request.headers.get('Upgrade')
+
+    # All on_xxx methods are in use by httptools parser.
+    # See https://github.com/MagicStack/httptools#apis
+    def on_header(self, name: bytes, value: bytes):
+        self.request.headers[name.decode().upper()] = value.decode()
+
+    def on_body(self, body: bytes):
+        # FIXME do not put all body in RAM blindly.
+        self.request.body += body
+
+    def on_url(self, url: bytes):
+        self.request.url = url
+        parsed = parse_url(url)
+        self.request.path = unquote(parsed.path.decode())
+        self.request.query_string = (parsed.query or b'').decode()
+
+    def on_message_begin(self):
+        self.request = self.app.Request(self.app, self.writer)
+        self.response = self.app.Response(self.app)
+
+    def on_message_complete(self):
+        self.request.method = self.parser.get_method().decode().upper()
+        task = self.app.loop.create_task(self.app(self.request, self.response))
+        task.add_done_callback(self.write)
+
+    # May or may not have "future" as arg.
+    def write(self, *args):
+        if self.writer.is_closing():
+            # We can't write, the writer is closing or is closed.
+            return
+
+        # Appends bytes for performances.
+        payload = b'HTTP/1.1 %a %b\r\n' % (
+            self.response.status.value, self.response.status.phrase.encode())
+        if not isinstance(self.response.body, bytes):
+            self.response.body = str(self.response.body).encode()
+        # https://tools.ietf.org/html/rfc7230#section-3.3.2 :scream:
+        bodyless = (self.response.status in self._BODYLESS_STATUSES or
+                    (hasattr(self, 'request') and
+                     self.request.method in self._BODYLESS_METHODS))
+        if 'Content-Length' not in self.response.headers and not bodyless:
+            length = len(self.response.body)
+            self.response.headers['Content-Length'] = length
+        if self.response._cookies:
+            # https://tools.ietf.org/html/rfc7230#page-23
+            for cookie in self.response.cookies.values():
+                payload += b'Set-Cookie: %b\r\n' % str(cookie).encode()
+        for key, value in self.response.headers.items():
+            payload += b'%b: %b\r\n' % (key.encode(), str(value).encode())
+        payload += b'\r\n'
+        if self.response.body and not bodyless:
+            payload += self.response.body
+        self.writer.write(payload)
+        if not self.parser.should_keep_alive():
+            self.writer.close()
+
+    
 Route = namedtuple('Route', ['payload', 'vars'])
+Protocols = {
+    'websocket': WebsocketHandler,
+}
 
 
-class Roll:
+class Roll(dict):
     """Deal with routes dispatching and events listening.
 
     You can subclass it to set your own `Protocol`, `Routes`, `Query`, `Form`,
@@ -368,11 +463,19 @@ class Roll:
     def factory(self):
         return self.Protocol(self)
 
-    def route(self, path: str, methods: list=None, **extras: dict):
+    def route(self, path: str, methods: list=None,
+              protocol: str='http', **extras: dict):
+
         if methods is None:
             methods = ['GET']
+        elif protocol == 'websocket' and methods != ['GET']:
+            raise RuntimeError('Websockets can only be registered on GET.')
 
         def wrapper(func):
+            proxy = Protocols.get(protocol, None)
+            if proxy is not None:
+                func = proxy(func)
+
             payload = {method: func for method in methods}
             payload.update(extras)
             self.routes.add(path, **payload)
@@ -395,70 +498,3 @@ class Roll:
         except KeyError:
             # Nobody registered to this event, let's roll anyway.
             pass
-
-
-class WSRoll(Roll):
-
-    Protocol = WSProtocol
-
-    def __init__(self):
-        super().__init__()
-        self.websockets = set()  # set of 2 items tuple, (task, websocket)
-
-    async def on_error(self, request: Request, response: Response, error):
-        if request.route.payload['is_websocket'] is None:
-            # This is not a websocket.
-            # Report the HTTP error as planned.
-            return await super().on_error(request, response, error)
-        print_exc()
-
-    def route(self, path: str, websocket: bool=False,
-              subprotocols: list=None, methods: list=None, **extras: dict):
-
-        if not websocket:
-            return super(WSRoll, self).route(path, methods=methods, **extras)
-
-        if methods and methods != ['GET']:
-            raise RuntimeError('Websockets can only handshake on GET')
-
-        extras['is_websocket'] = websocket
-        if subprotocols:
-            subprotocols = frozenset(subprotocols)  # Set in stone.
-        def ws_wrapper(func):
-            async def websocket_handler(request, response, **params):
-                protocol = request.transport.get_protocol()
-                ws = protocol.websocket_handshake(request, subprotocols)
-                fut = asyncio.ensure_future(
-                    func(request, ws, **params), loop=self.loop)
-                self.websockets.add((fut, ws))
-                try:
-                    await fut
-                except ConnectionClosed:
-                    # The client closed the connection.
-                    # We cancel the future to be sure it's in order.
-                    fut.cancel()
-                    await ws.close(1002, 'Connection closed untimely.')
-                except asyncio.CancelledError:
-                    # The websocket task was cancelled
-                    # We need to warn the client.
-                    await ws.close(1001, 'Handler cancelled.')
-                except Exception as exc:
-                    # A more serious error happened.
-                    # The websocket handler was untimely terminated
-                    # by an unwarranted exception. Warn the client.
-                    await ws.close(1011, 'Handler died prematurely.')
-                    raise
-                else:
-                    # The handler finished gracefully.
-                    # We can close the socket in peace.
-                    await ws.close()
-                finally:
-                    # Whatever happened, the websocket fate has been
-                    # sealed. We remove it from our watch.
-                    self.websockets.discard((fut, ws))
-
-            payload = {'GET':  websocket_handler}
-            payload.update(extras)
-            self.routes.add(path, **payload)
-
-        return ws_wrapper
