+++ tests/test_websockets.py
@@ -9,7 +9,7 @@ import websockets
 async def test_websocket_route(liveclient):
     ev = asyncio.Event()
 
-    @liveclient.app.route('/ws', websocket=True)
+    @liveclient.app.route('/ws', protocol="websocket")
     async def handler(request, ws, **params):
         assert ws.subprotocol is None
         ev.set()
@@ -23,18 +23,18 @@ async def test_websocket_route(liveclient):
     assert response.status_code == 101
 
     with pytest.raises(RuntimeError) as exc:
-        @liveclient.app.route('/ws', websocket=True, methods=['POST'])
+        @liveclient.app.route('/ws', protocol="websocket", methods=['POST'])
         async def handler(request, ws, **params):
             assert ws.subprotocol is None
             ev.set()
 
-    assert str(exc.value) == 'Websockets can only handshake on GET'
+    assert str(exc.value) == 'Websockets can only be registered on GET.'
 
 
 @pytest.mark.asyncio
 async def test_websocket_communication(liveclient):
 
-    @liveclient.app.route('/echo', websocket=True)
+    @liveclient.app.route('/echo', protocol="websocket")
     async def echo(request, ws, **params):
         while True:
             message = await ws.recv()
@@ -58,7 +58,7 @@ async def test_websocket_communication(liveclient):
 @pytest.mark.asyncio
 async def test_websocket_broadcasting(liveclient):
 
-    @liveclient.app.route('/broadcast', websocket=True)
+    @liveclient.app.route('/broadcast', protocol="websocket")
     async def feed(request, ws, **params):
         while True:
             message = await ws.recv()
@@ -89,7 +89,7 @@ async def test_websocket_broadcasting(liveclient):
 @pytest.mark.asyncio
 async def test_websocket_binary(liveclient):
 
-    @liveclient.app.route('/bin', websocket=True)
+    @liveclient.app.route('/bin', protocol="websocket")
     async def binary(request, ws, **params):
         await ws.send(b'test')
 
@@ -106,7 +106,7 @@ async def test_websocket_binary(liveclient):
 async def test_websocket_route_with_subprotocols(liveclient):
     results = []
 
-    @liveclient.app.route('/ws', websocket=True, subprotocols=['foo', 'bar'])
+    @liveclient.app.route('/ws', protocol="websocket", subprotocols=['foo', 'bar'])
     async def handler(request, ws):
         results.append(ws.subprotocol)
